# apps/web-v2/src/lib/generated/collections.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CurateDS.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/collections": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CurateDS.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collectionId}/summary": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collectionId}/insights": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "attributeDefinitionId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionInsightsDto"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collectionId}/activity": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedCollectionActivityResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CollectionResponse": {
        "required": [
          "id",
          "name",
          "createdUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateCollectionRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CollectionSummaryResponse": {
        "required": [
          "collectionId",
          "totalItems",
          "totalAttributeDefinitions",
          "tagsUsed",
          "locationsUsed",
          "itemsWithNoLocation",
          "itemsWithNoTags",
          "totalMediaAssets"
        ],
        "type": "object",
        "properties": {
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "totalItems": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalAttributeDefinitions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tagsUsed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "locationsUsed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "itemsWithNoLocation": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "itemsWithNoTags": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalMediaAssets": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CollectionInsightsDto": {
        "required": [
          "collectionId",
          "summary",
          "reports",
          "itemsByType",
          "addedByMonth",
          "attribute"
        ],
        "type": "object",
        "properties": {
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "summary": {
            "$ref": "#/components/schemas/CollectionSummaryDto"
          },
          "reports": {
            "$ref": "#/components/schemas/CollectionReportsDto"
          },
          "itemsByType": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemsByTypeDto"
            }
          },
          "addedByMonth": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemsAddedMonthDto"
            }
          },
          "attribute": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AttributeBreakdownDto"
              }
            ]
          }
        }
      },
      "CollectionSummaryDto": {
        "required": [
          "collectionId",
          "totalItems",
          "totalAttributeDefinitions",
          "tagsUsed",
          "locationsUsed",
          "itemsWithNoLocation",
          "itemsWithNoTags",
          "totalMediaAssets"
        ],
        "type": "object",
        "properties": {
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "totalItems": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalAttributeDefinitions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tagsUsed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "locationsUsed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "itemsWithNoLocation": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "itemsWithNoTags": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalMediaAssets": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CollectionReportsDto": {
        "required": [
          "itemsByLocation",
          "itemsByTag"
        ],
        "type": "object",
        "properties": {
          "itemsByLocation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemsByLocationDto"
            }
          },
          "itemsByTag": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemsByTagDto"
            }
          }
        }
      },
      "ItemsByLocationDto": {
        "required": [
          "locationId",
          "locationName",
          "count"
        ],
        "type": "object",
        "properties": {
          "locationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "locationName": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ItemsByTagDto": {
        "required": [
          "tagId",
          "tagName",
          "count"
        ],
        "type": "object",
        "properties": {
          "tagId": {
            "type": "string",
            "format": "uuid"
          },
          "tagName": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ItemsByTypeDto": {
        "required": [
          "itemTypeId",
          "name",
          "count"
        ],
        "type": "object",
        "properties": {
          "itemTypeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ItemsAddedMonthDto": {
        "required": [
          "fromUtc",
          "toUtc",
          "count"
        ],
        "type": "object",
        "properties": {
          "fromUtc": {
            "type": "string",
            "format": "date-time"
          },
          "toUtc": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "AttributeBreakdownDto": {
        "required": [
          "definitionId",
          "key",
          "name",
          "totalWithValue",
          "values"
        ],
        "type": "object",
        "properties": {
          "definitionId": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "totalWithValue": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttributeBucketDto"
            }
          }
        }
      },
      "AttributeBucketDto": {
        "required": [
          "value",
          "count"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "PagedCollectionActivityResponse": {
        "required": [
          "events",
          "totalCount",
          "page",
          "pageSize",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionActivityEventResponse"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CollectionActivityEventResponse": {
        "required": [
          "eventId",
          "itemId",
          "itemName",
          "eventType",
          "occurredUtc",
          "occurredBy",
          "notes"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "itemName": {
            "type": "string"
          },
          "eventType": {
            "type": "string"
          },
          "occurredUtc": {
            "type": "string",
            "format": "date-time"
          },
          "occurredBy": {
            "type": "string"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  }
}